@@ -1,6 +1,6 @@
 {
-  "name": "vue-konva",
-  "version": "2.1.7",
+  "name": "vue3-konva",
+  "version": "1.0.0",
   "description": "Vue binding to canvas element via Konva framework",
   "keywords": [
     "vue",
@@ -53,13 +53,13 @@
     "konva": ">3.2.5"
   },
   "devDependencies": {
-    "@vue/test-utils": "^1.0.0-beta.25",
+    "@vue/compiler-sfc": "^3.0.4",
+    "@vue/test-utils": "^2.0.0-beta.10",
     "chai": "^4.2.0",
     "konva": ">3.2.5",
     "nwb": "^0.25.2",
     "sinon": "^7.3.0",
-    "vue": "^2.6.11",
-    "vue-template-compiler": "^2.6.11"
+    "vue": "^3.0.11"
   },
   "engines": {
     "node": ">= 4.0.0",
